backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-filter>=25.2",
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
